переделывает модули
