+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Кекстаграм (простой)»
 
 * Студент: [Сергей Курдюков](https://up.htmlacademy.ru/javascript/27/user/2027773).
-* Наставник: Roman Zheltov .
+* Наставник: Виктория Никитина.
 
 ---
