+++ js/main.js
@@ -1,1 +1,4 @@
 import './data.js';
+import { getData } from './data';
+
+getData();
