+++ js/utilits.js
@@ -8,3 +8,14 @@ function searchRandomNumber (min, max) {
 }
 
 export { searchRandomNumber };
+
+function checksLength ( str, length ) {
+  if (typeof str === 'string') {
+    return str.length <= length;
+  }
+  return NaN;
+}
+
+checksLength ();
+
+export { checksLength };
